execution/state: move the log tests next to the arena, note EIP-7825
The log tests and benchmarks lived in two files named after IntraBlockState;
they belong with the type they exercise, so they move to log_arena_test.go.

The budget comment also carried 45M gas limit figures — mainnet is at 60M —
and said nothing about the per-transaction cap, which is the ceiling that
matters when the arena resets once per transaction.
