+++ execution/state/log_arena_test.go
@@ -428,3 +428,184 @@ func BenchmarkLogsRlpHash(b *testing.B) {
 		}
 	})
 }
+
+// TestAddLogOnLogHookCopyContract pins the OnLog contract on the AddLog path:
+// the hook sees the emitted contents during the callback, and a copy taken then
+// stays intact after the buffer entry is reused by later blocks.
+func TestAddLogOnLogHookCopyContract(t *testing.T) {
+	t.Parallel()
+
+	ibs := New(NewNoopReader())
+	var copied []*types.Log
+	ibs.SetHooks(&tracing.Hooks{
+		OnLog: func(l *types.Log) { copied = append(copied, l.Copy()) },
+	})
+
+	ibs.SetTxContext(1, 0)
+	ibs.AddLog(&types.Log{Address: common.Address{0x11}, Topics: []common.Hash{{0x01}}, Data: []byte{0x01}})
+	require.Len(t, copied, 1)
+
+	for i := range 1000 {
+		ibs.AddLog(&types.Log{Address: common.Address{0x22}, Data: []byte{byte(i)}})
+	}
+	ibs.Reset()
+	ibs.SetTxContext(2, 0)
+	ibs.AddLog(&types.Log{Address: common.Address{0x33}, Data: []byte{0x99}})
+
+	require.Equal(t, common.Address{0x11}, copied[0].Address)
+	require.Equal(t, []common.Hash{{0x01}}, copied[0].Topics)
+	require.Equal(t, []byte{0x01}, []byte(copied[0].Data))
+}
+
+// TestNotifyLogHookGetsStableCopy pins the OnLog contract on the AllocLog path
+// the EVM's makeLog uses: the hook owns what it receives, so retaining it is
+// safe even though the buffer entry behind it is reused by a later block.
+func TestNotifyLogHookGetsStableCopy(t *testing.T) {
+	t.Parallel()
+
+	ibs := New(NewNoopReader())
+	var handed []*types.Log
+	ibs.SetHooks(&tracing.Hooks{
+		OnLog: func(l *types.Log) { handed = append(handed, l) },
+	})
+
+	emit := func(addr common.Address, topic common.Hash, data []byte) *types.Log {
+		lp := ibs.AllocLog(addr, 1, len(data))
+		lp.Topics[0] = topic
+		copy(lp.Data, data)
+		ibs.NotifyLog(lp)
+		return lp
+	}
+
+	ibs.SetTxContext(1, 0)
+	lp := emit(common.Address{0xaa}, common.Hash{0x01}, []byte{0x11, 0x22})
+	require.NotSame(t, lp, handed[0])
+
+	ibs.Reset()
+	ibs.SetTxContext(2, 0)
+	require.Same(t, lp, emit(common.Address{0xbb}, common.Hash{0x99}, []byte{0xde, 0xad}))
+
+	require.Equal(t, common.Hash{0x01}, handed[0].Topics[0])
+	require.Equal(t, []byte{0x11, 0x22}, []byte(handed[0].Data))
+	require.Equal(t, common.Hash{0x99}, handed[1].Topics[0])
+	require.Equal(t, []byte{0xde, 0xad}, []byte(handed[1].Data))
+}
+
+// TestAddLogKeepsCallerBlockNumber pins that BlockNumber stays a caller-assigned
+// field: execution/state does not know the block number, so a reused entry must
+// not carry the previous caller's value either.
+func TestAddLogKeepsCallerBlockNumber(t *testing.T) {
+	t.Parallel()
+
+	ibs := New(NewNoopReader())
+	ibs.SetTxContext(7, 0)
+	ibs.AddLog(&types.Log{Address: common.Address{0x01}, BlockNumber: 42})
+	ibs.AddLog(&types.Log{Address: common.Address{0x02}})
+
+	logs := ibs.GetRawLogs(0)
+	require.Len(t, logs, 2)
+	require.Equal(t, hexutil.Uint64(42), logs[0].BlockNumber)
+	require.Zero(t, logs[1].BlockNumber, "an unset BlockNumber must not inherit the previous log's")
+}
+
+// TestAllocLogPreservesCapacityAcrossRevert pins that fully reverting a tx's
+// logs (which truncates the outer buffer) and then logging again reuses the
+// inner buffer's capacity instead of dropping it — the same capacity Reset
+// preserves.
+func TestAllocLogPreservesCapacityAcrossRevert(t *testing.T) {
+	t.Parallel()
+
+	ibs := New(NewNoopReader())
+	ibs.SetTxContext(1, 0)
+	snap := ibs.PushSnapshot()
+	for i := range 8 {
+		ibs.AddLog(&types.Log{Address: common.Address{byte(i)}})
+	}
+	require.Len(t, ibs.logs.byTx, 2)
+	capBefore := cap(ibs.logs.byTx[1])
+	require.GreaterOrEqual(t, capBefore, 8)
+
+	ibs.RevertToSnapshot(snap, nil)
+	require.Len(t, ibs.logs.byTx, 1) // the tx's slot was truncated off the outer buffer
+
+	ibs.AddLog(&types.Log{Address: common.Address{0xff}})
+	require.Len(t, ibs.logs.byTx, 2)
+	require.Equal(t, capBefore, cap(ibs.logs.byTx[1]), "inner log buffer capacity must survive revert+relog")
+}
+
+// A block of large logs must not pay for them again next block: inside the
+// budget an oversized buffer is kept, like any other.
+func TestResetKeepsOversizedLogDataWithinBudget(t *testing.T) {
+	t.Parallel()
+
+	ibs := New(NewNoopReader())
+	ibs.SetTxContext(1, 0)
+	small := ibs.AllocLog(common.Address{0x01}, 1, 8)
+	big := ibs.AllocLog(common.Address{0x02}, 1, maxReusableLogDataCap+1)
+	bigData := big.Data[:1]
+
+	ibs.Reset()
+	ibs.SetTxContext(2, 0)
+	require.Same(t, small, ibs.AllocLog(common.Address{0x03}, 1, 8), "normal-size entry is reused")
+	relog := ibs.AllocLog(common.Address{0x04}, 1, maxReusableLogDataCap+1)
+	require.Same(t, big, relog, "the entry is reused")
+	require.Equal(t, &bigData[0], &relog.Data[0], "and so is its Data buffer")
+}
+
+// Past the byte budget the oversized buffers are what goes, and the newest go
+// first: transactions take positions in order, so the head is what the next
+// block asks for first.
+func TestResetEvictsOversizedLogDataFirst(t *testing.T) {
+	t.Parallel()
+
+	const dataSize = maxReusableLogDataCap + 1
+	txs := maxReusableLogBytes/dataSize + 2
+
+	ibs := New(NewNoopReader())
+	for txIndex := range txs {
+		ibs.SetTxContext(1, txIndex)
+		ibs.AllocLog(common.Address{0x01}, 0, dataSize)
+	}
+	ibs.Reset()
+
+	_, _, dataBytes := retainedLogs(ibs)
+	require.LessOrEqual(t, dataBytes, maxReusableLogBytes)
+	require.Nil(t, entryAt(ibs, txs, 0).Data, "the newest buffer goes first")
+	require.NotNil(t, entryAt(ibs, 1, 0).Data, "the head stays warm for the next block")
+}
+
+// TestLogIndexIsBlockWide pins that AddLog stamps a block-wide log index:
+// receipts.DeriveFields derives FirstLogIndexWithinBlock from Logs[0].Index, so
+// the counter must run across transactions, roll back with a reverted log, and
+// restart at zero on Reset.
+func TestLogIndexIsBlockWide(t *testing.T) {
+	t.Parallel()
+
+	ibs := New(NewNoopReader())
+	ibs.SetTxContext(1, 0)
+	ibs.AddLog(&types.Log{Address: common.Address{0x01}})
+	ibs.AddLog(&types.Log{Address: common.Address{0x02}})
+
+	ibs.SetTxContext(1, 1)
+	snap := ibs.PushSnapshot()
+	ibs.AddLog(&types.Log{Address: common.Address{0x03}})
+	ibs.RevertToSnapshot(snap, nil)
+	ibs.AddLog(&types.Log{Address: common.Address{0x04}})
+
+	tx0, tx1 := ibs.GetRawLogs(0), ibs.GetRawLogs(1)
+	require.Len(t, tx0, 2)
+	require.Len(t, tx1, 1)
+	require.Equal(t, hexutil.Uint(0), tx0[0].Index)
+	require.Equal(t, hexutil.Uint(1), tx0[1].Index)
+	require.Equal(t, hexutil.Uint(2), tx1[0].Index, "index continues across txs and reuses a reverted slot")
+
+	ibs.Reset()
+	ibs.SetTxContext(2, 0)
+	ibs.AddLog(&types.Log{Address: common.Address{0x05}})
+	require.Equal(t, hexutil.Uint(0), ibs.GetRawLogs(0)[0].Index, "next block restarts at zero")
+}
+
+func entryAt(ibs *IntraBlockState, ti, idx int) *types.Log {
+	entries := ibs.logs.byTx[ti]
+	return entries[:cap(entries)][idx]
+}
